fix: Do not set any model defaults when forwarding the requests
Different versions of Gemini have different defaults for maximum
output tokens. The default 1024 in the code was not enough when
asking Gemini 2.5 reegular questions.

BREAKING CHANGE: There were default values for the Gemini models
in the code: `maxOutputTokens: 1024` and temperature: 0.5`. They
were removed in favour of the Gemini defaults as implemented
in the model itself. If you depended on the specific values,
set them explicitely in `generationConfig` in `model-defaults.json`
or in `options` when making the request to the Ollama REST API.